requirements.txt
absl-py==2.1.0
asttokens==3.0.0
astunparse==1.6.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.1.0
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.0
h5py==3.12.1
idna==3.10
ipython==8.30.0
ipython-autotime==0.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
keras==3.7.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.48
protobuf==5.29.0
pure_eval==0.2.3
Pygments==2.18.0
requests==2.32.3
rich==13.9.4
scipy==1.14.1
six==1.16.0
split-folders==0.5.1
stack-data==0.6.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
